+++ autohyperlink-urls.php
@@ -10,7 +10,7 @@
  * License URI: https://www.gnu.org/licenses/gpl-2.0.html
  * Description: Automatically turns plaintext URLs and email addresses into links.
  *
- * Compatible with WordPress 4.7 through 5.1+.
+ * Compatible with WordPress 4.7 through 5.3+.
  *
  * =>> Read the accompanying readme.txt file for instructions and documentation.
  * =>> Also, visit the plugin's homepage for additional information and updates.