+++ autohyperlink-urls.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * Plugin Name: Auto-hyperlink URLs
- * Version:     5.3
+ * Version:     5.4
  * Plugin URI:  http://coffee2code.com/wp-plugins/auto-hyperlink-urls/
  * Author:      Scott Reilly
  * Author URI:  http://coffee2code.com/
@@ -18,7 +18,7 @@
  *
  * @package Auto_Hyperlink_URLs
  * @author  Scott Reilly
- * @version 5.3
+ * @version 5.4
  */
 
 /*
@@ -110,7 +110,7 @@ final class c2c_AutoHyperlinkURLs extends c2c_AutoHyperlinkURLs_Plugin_049 {
 	 * Constructor.
 	 */
 	protected function __construct() {
-		parent::__construct( '5.3', 'autohyperlink-urls', 'c2c', __FILE__, array() );
+		parent::__construct( '5.4', 'autohyperlink-urls', 'c2c', __FILE__, array() );
 		register_activation_hook( __FILE__, array( __CLASS__, 'activation' ) );
 
 		return self::$instance = $this;