+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     * Autolinks the following ACF field types: text, textarea, url, wysiwyg
     * Adds filter `c2c_autohyperlink_acf_urls_filters` for customizing which ACF-related filters to hook
 * Change: Note compatibility through WP 5.3+
+* Change: Minor tweaks to descriptions of functions in inline documentation
 * Change: Update copyright date (2020)
 
 ## 5.3 _(2019-04-19)_