+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 * New: Add filter `autohyperlink_no_autolink_content_tags` for configuring which HTML tags don't get their content autolinked
 * New: Add support for Advanced Custom Fields (ACF) plugin fields
     * Autolinks the following ACF field types: text, textarea, url, wysiwyg
-    * Adds filter `c2c_autohyperlink_acf_urls_filters` for customizing which ACF-related filters to hook
+    * Adds filter `c2c_autohyperlink_urls_acf_filters` for customizing which ACF-related filters to hook
 * New: Unit tests: Add tests to verify default hooks get hooked
 * Change: Note compatibility through WP 5.3+
 * Change: Minor tweaks to descriptions of functions in inline documentation